// src/interface/approve-deliver/i-req-mpls_create_send_car_deliver_and_loyalty_consent.interface.ts
export interface IReqMplsCreateSendCarDeliverAndLoyaltyConsent {
    is_check_sale_sheet_explain: number | null
    is_check_product_detail: number | null
    is_check_payment_rule: number | null
    is_check_contract_explain: number | null
    is_check_total_loss_explain: number | null
    is_check_total_loss_company: number | null
    lalon: string
    latitude: string
    londtiude: string
}